# tests/golden/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "EdgeRecord": {
        "description": "One typed relationship.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "dst_id": {
            "title": "Dst Id",
            "type": "integer"
          },
          "ended_at": {
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "rel": {
            "title": "Rel",
            "type": "string"
          },
          "src_id": {
            "title": "Src Id",
            "type": "integer"
          },
          "started_at": {
            "title": "Started At"
          }
        },
        "required": [
          "id",
          "src_id",
          "rel",
          "dst_id",
          "confidence",
          "started_at",
          "ended_at",
          "note"
        ],
        "title": "EdgeRecord",
        "type": "object"
      },
      "EntityKind": {
        "description": "The eight entity kinds (spec §4.1).",
        "enum": [
          "skill",
          "project",
          "organization",
          "role",
          "education",
          "credential",
          "achievement",
          "responsibility"
        ],
        "title": "EntityKind",
        "type": "string"
      },
      "EntityRecord": {
        "description": "One entity with its kind-specific attributes.",
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/EntityKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "normalized_name": {
            "title": "Normalized Name",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/KnowledgeState"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "normalized_name",
          "summary",
          "state",
          "first_seen_at",
          "updated_at"
        ],
        "title": "EntityRecord",
        "type": "object"
      },
      "EvidenceRecord": {
        "description": "One evidence span resolved to its source.\n\nAttributes:\n    id: The evidence id.\n    source_id: The source it was cut from.\n    source_kind: The source's kind.\n    source_label: The source's human label.\n    locator: Where in the source the span sits.\n    excerpt: The literal text of the span.",
        "properties": {
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_kind",
          "source_label",
          "locator",
          "excerpt"
        ],
        "title": "EvidenceRecord",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeState": {
        "description": "Knowledge state of an entity (spec §4.4). Not a match verdict.",
        "enum": [
          "confirmed",
          "learning",
          "gap",
          "archived"
        ],
        "title": "KnowledgeState",
        "type": "string"
      },
...
      "ProvenanceRecord": {
        "description": "One assertion resolved back to the literal evidence it cites (FR-03).",
        "properties": {
          "assertion_id": {
            "title": "Assertion Id",
            "type": "integer"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "assertion_id",
          "field",
          "value",
          "source_id",
          "source_kind",
          "source_label",
          "locator",
          "excerpt"
        ],
        "title": "ProvenanceRecord",
        "type": "object"
      },
...
      "TargetKind": {
        "description": "What an assertion or a classification points at.",
        "enum": [
          "entity",
          "edge"
        ],
        "title": "TargetKind",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CVForge"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/entities": {
      "get": {
        "description": "List entities, optionally filtered by kind and knowledge state.\n\nArgs:\n    conn: Read connection.\n    kind: Only this kind.\n    state: Only this knowledge state.\n\nReturns:\n    Matching entities.",
        "operationId": "list_entities_api_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EntityKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntityRecord"
                  },
                  "title": "Response List Entities Api Entities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Entities",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/entities/{entity_id}": {
      "get": {
        "description": "Return one entity with its kind-specific attributes.\n\nArgs:\n    conn: Read connection.\n    entity_id: The entity id.\n\nReturns:\n    The entity.\n\nRaises:\n    HTTPException: 404 if it does not exist.",
        "operationId": "get_entity_api_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entity",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/entities/{entity_id}/edges": {
      "get": {
        "description": "Return every edge touching an entity.\n\nArgs:\n    conn: Read connection.\n    entity_id: The entity id.\n\nReturns:\n    The edges.",
        "operationId": "entity_edges_api_entities__entity_id__edges_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EdgeRecord"
                  },
                  "title": "Response Entity Edges Api Entities  Entity Id  Edges Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entity Edges",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/evidence/{evidence_id}": {
      "get": {
        "description": "Return an evidence span's literal excerpt and its source (FR-03).\n\nArgs:\n    conn: Read connection.\n    evidence_id: The evidence id an operation payload carries.\n\nReturns:\n    The evidence record.\n\nRaises:\n    HTTPException: 404 if the evidence does not exist.",
        "operationId": "get_evidence_api_evidence__evidence_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "knowledge"
        ]
      }
    },
...
    "/api/provenance/{target_kind}/{target_id}": {
      "get": {
        "description": "Show where a stored fact came from: source kind, locator and excerpt (FR-03).\n\nArgs:\n    conn: Read connection.\n    target_kind: ``entity`` or ``edge``.\n    target_id: Its id.\n\nReturns:\n    One record per assertion.\n\nRaises:\n    HTTPException: 404 if nothing is asserted about that target.",
        "operationId": "provenance_api_provenance__target_kind___target_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "target_kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TargetKind"
            }
          },
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProvenanceRecord"
                  },
                  "title": "Response Provenance Api Provenance  Target Kind   Target Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provenance",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}